requirements.txt
fastapi===0.111.0
uvicorn===0.30.1
asyncpraw===7.7.1
StackAPI===0.3.1
loguru===0.7.2
Markdown==3.6
better-profanity===0.7.0
google-generativeai===0.6.0